+++ imports/ui/pages/Customer.jsx
@@ -11,6 +11,7 @@ import { Customers } from '../../api/customers/customersCollection';
 import { Quotes } from '../../api/quotes/quotesCollection';
 
 import EditCustomerConnect from '../editors/EditCustomerConnect';
+import EditShipmentConnect from '../editors/EditShipmentConnect.jsx';
 import EditQuoteHeaderConnect from '../editors/EditQuoteHeaderConnect.jsx';
 import EditQuoteChargesConnect from '../editors/EditQuoteChargesConnect';
 import EditQuoteEmailConnect from '../editors/EditQuoteEmailConnect';
@@ -74,6 +75,10 @@ const CustomerInner = ({ customer, loading, history }) => {
         <Route
           path="/customers/view/:customerId/quotes/:quoteId/view"
           render={props => <ViewQuote {...props} />}
+        />
+        <Route
+          path="/customers/view/:customerId/shipments/:shipmentId"
+          render={props => <EditShipmentConnect {...props} />}
         />
       </div>
       <div className="content-footer-accent customers-footer-accent" />